Completely removed the Models folder from WebApi project
Renamed Web.Api namespace from the project configuration to stop Resharper from giving warnings about type namespaces not corresponding to their physical location
