+++ Procultura.Application/DTO/User/LoginModel.cs
@@ -1,0 +1,10 @@
+﻿namespace Procultura.Application.DTO.User
+{
+    using Procultura.Application.DTO;
+
+    public class LoginModel: RequestBase
+    {
+        public string Email { get; set; }
+        public string Password { get; set; }
+    }
+}
